+++ src/main/java/study/board/Service/CommentService.java
@@ -9,6 +9,8 @@ import study.board.dto.request.CommentUpdateRequestDto;
 import study.board.entity.Comment;
 import study.board.entity.Member;
 import study.board.entity.Post;
+import study.board.entity.Role;
+import study.board.exception.AuthorizationFailException;
 import study.board.exception.NoCommentException;
 import study.board.exception.NoPostException;
 import study.board.exception.NoUserException;
@@ -17,6 +19,8 @@ import study.board.repository.member.MemberRepository;
 import study.board.repository.post.PostRepository;
 
 import java.util.List;
+
+import static study.board.entity.Role.*;
 
 @Service
 @RequiredArgsConstructor
@@ -42,14 +46,22 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(CommentUpdateRequestDto dto, long commentId) {
+    public void updateComment(CommentUpdateRequestDto dto, long commentId, Member member) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(NoCommentException::new);
-        comment.update(dto);
+        if (comment.getMember() == member || member.getRole() == ADMIN) {
+            comment.update(dto);
+        } else {
+            throw new AuthorizationFailException();
+        }
     }
 
     @Transactional
-    public void deleteComment(long commentId) {
+    public void deleteComment(long commentId, Member member) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(NoCommentException::new);
-        commentRepository.delete(comment);
+        if (comment.getMember() == member || member.getRole() == ADMIN) {
+            commentRepository.delete(comment);
+        } else {
+            throw new AuthorizationFailException();
+        }
     }
 }
